Fix GitHub issue: Files showing as modified directly after git clone (checkout new files)
In Visual Studio, if you are using Git, you can auto generate the
.gitignore and .gitattributes files. The auto generated .getattributes
file has the following line:

* text=auto

This line is near the top of the file. We just needed to comment the
line out by added a # to the front of it. After doing that, things
operated as expected